move BlockMeta class to its own file
