+++ src/main/java/com/tmtravlr/soundfilters/SoundFiltersMod.java
@@ -1,12 +1,10 @@
 package com.tmtravlr.soundfilters;
 
-import static com.tmtravlr.soundfilters.SoundFiltersConfig.ALL_METAS;
 import static com.tmtravlr.soundfilters.SoundFiltersConfig.config;
 import static com.tmtravlr.soundfilters.SoundFiltersConfig.configure;
 import static com.tmtravlr.soundfilters.SoundFiltersConfig.logger;
 import static com.tmtravlr.soundfilters.SoundFiltersConfig.reverbFilter;
 
-import net.minecraft.block.Block;
 import net.minecraftforge.common.config.Configuration;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.common.Mod.EventHandler;
@@ -49,56 +47,6 @@ public class SoundFiltersMod {
 		}
 
 		System.out.println("[Sound Filters] Loaded modified library.");
-	}
-
-	/**
-	 * Holds info about both a block's id and metadata.
-	 *
-	 * @author Rebeca
-	 * @Date 2014
-	 */
-	public static class BlockMeta {
-		public Block block;
-		public int meta;
-
-		public BlockMeta(Block b, int m) {
-			block = b;
-			meta = m;
-		}
-
-		@Override
-		public int hashCode() {
-			final int prime = 31;
-			int result = 1;
-			result = prime * result + Block.getIdFromBlock(block);
-			result = prime * result + meta;
-			return result;
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			if (this == obj) { return true; }
-			if (!(obj instanceof BlockMeta)) { return false; }
-			BlockMeta other = (BlockMeta) obj;
-			if (block != other.block && Block.getIdFromBlock(block) != Block.getIdFromBlock(other.block)) return false;
-			if (meta != other.meta) { return false; }
-			return true;
-		}
-
-		@Override
-		public String toString() {
-			StringBuilder builder = new StringBuilder();
-			builder.append("BlockMeta [block=");
-			builder.append(block);
-			builder.append(", meta=");
-			if (meta == ALL_METAS) {
-				builder.append("all");
-			} else {
-				builder.append(meta);
-			}
-			builder.append(']');
-			return builder.toString();
-		}
 	}
 
 	@EventHandler
